+++ packages/editor/editor-core/example-helpers/quick-insert-provider.tsx
@@ -83,6 +83,48 @@ const items: Array<QuickInsertItem> = [
     },
   },
   {
+    title: 'Minimum width extension',
+    icon: () => <DevIcon label="dev" />,
+    action(insert) {
+      return insert({
+        type: 'extension',
+        attrs: {
+          extensionType: 'com.atlassian.confluence.macro.core',
+          extensionKey: 'block-layout-eh',
+          text: 'Minimum width block extension demo',
+          parameters: {
+            style: { minWidth: 400 },
+          },
+        },
+      });
+    },
+  },
+  {
+    title: 'iframe Block extension',
+    icon: () => <DevIcon label="dev" />,
+    action(insert) {
+      return insert({
+        type: 'extension',
+        attrs: {
+          extensionType: 'com.atlassian.confluence.macro.core',
+          extensionKey: 'block-iframe-eh',
+          text: 'Full width block extension demo',
+          parameters: {
+            macroParams: {},
+            macroMetadata: {
+              placeholder: [
+                {
+                  data: { url: '' },
+                  type: 'icon',
+                },
+              ],
+            },
+          },
+        },
+      });
+    },
+  },
+  {
     title: 'Lorem ipsum',
     icon: () => <DevIcon label="dev" />,
     action(insert) {