+++ packages/elements/mention/src/__tests__/unit/api/TeamMentionResourceSpec.ts
@@ -46,20 +46,43 @@ const FULL_CONTEXT = {
   childObjectId: 'someChildObjectId',
   sessionId: 'someSessionId',
 };
+const REQUEST_DELAY = 100;
 
 describe('TeamMentionResourceSpec', () => {
   let resource: TeamMentionResource;
+  let currentCount = 0;
+  const doneWhen = (done: () => void, maxCount: number) => {
+    if (currentCount === maxCount) {
+      done();
+    }
+    ++currentCount;
+  };
 
   beforeEach(() => {
+    jest.useFakeTimers();
+
+    currentCount = 1;
+
+    resource = new TeamMentionResource(
+      apiUserMentionConfig,
+      apiTeamMentionConfig,
+    );
+
     fetchMock
       .mock(/\/users\/mentions\/search\?.*query=craig(&|$)/, {
         body: {
           mentions: resultCraig,
         },
       })
-      .mock(/\/teams\/mentions\/search\?.*query=craig(&|$)/, {
-        body: teamResults,
-      })
+      .mock(
+        /\/teams\/mentions\/search\?.*query=craig(&|$)/,
+        // assume team request is slow
+        new Promise(resolve => {
+          setTimeout(() => {
+            resolve({ body: teamResults });
+          }, REQUEST_DELAY);
+        }),
+      )
       .mock(/\/users\/mentions\/search\?.*query=esoares(&|$)/, {
         body: {
           mentions: [],
@@ -71,14 +94,14 @@ describe('TeamMentionResourceSpec', () => {
       .mock(
         /\/users\/mentions\/search\?.*query=cr(&|$)/,
         new Promise(resolve => {
-          window.setTimeout(() => {
+          setTimeout(() => {
             resolve({
               // delayed results
               body: {
                 mentions: resultCr,
               },
             });
-          }, 100);
+          }, REQUEST_DELAY);
         }),
       )
       .mock(/\/teams\/mentions\/search\?.*query=cr(&|$)/, {
@@ -98,25 +121,44 @@ describe('TeamMentionResourceSpec', () => {
         500,
       )
       .mock(/\/teams\/mentions\/search\?.*query=query-only-users-fail(&|$)/, {
-        body: {
-          mentions: teamResults,
-        },
+        body: teamResults,
+      })
+      .mock(/\/users\/mentions\/search\?.*query=query-both-fail(&|$)/, 500)
+      .mock(/\/teams\/mentions\/search\?.*query=query-both-fail(&|$)/, 500)
+      .mock(
+        /\/users\/mentions\/search\?.*team-faster-user(&|$)/,
+        // user request is slower than team request
+        new Promise(resolve => {
+          setTimeout(() => {
+            resolve({ body: { mentions: resultCraig } });
+          }, REQUEST_DELAY);
+        }),
+      )
+      .mock(/\/teams\/mentions\/search\?.*team-faster-user(&|$)/, {
+        body: teamResults,
       });
-
-    resource = new TeamMentionResource(
-      apiUserMentionConfig,
-      apiTeamMentionConfig,
-    );
   });
 
   afterEach(() => {
     fetchMock.restore();
+    jest.useRealTimers();
   });
 
   describe('#subscribe', () => {
-    it('subscribe should receive updates', done => {
+    it('subscribe should receive 2 updates: from user request first and then from team request', done => {
+      // the event is called twice
+      // 1st: with user results
+      // 2nd: with users and team results
       resource.subscribe('craig', mentions => {
-        expect(mentions).toHaveLength(resultCraig.length + teamResults.length);
+        // the first is for user results
+        if (currentCount === 1) {
+          expect(mentions).toHaveLength(resultCraig.length);
+          // the second is for user results + team results
+        } else if (currentCount === 2) {
+          expect(mentions).toHaveLength(
+            resultCraig.length + teamResults.length,
+          );
+        }
 
         const queryParams = queryString.parse(
           queryString.extract(fetchMock.lastUrl()),
@@ -128,34 +170,21 @@ describe('TeamMentionResourceSpec', () => {
         expect(queryParams.sessionId).toBe('someSessionId');
         expect(fetchMock.lastOptions().credentials).toEqual('include');
 
-        done();
+        doneWhen(done, 2);
       });
 
       resource.filter('craig', FULL_CONTEXT);
+      jest.runTimersToTime(REQUEST_DELAY);
     });
 
-    it('multiple subscriptions should receive updates', done => {
-      let count = 0;
-
-      resource.subscribe('test1', mentions => {
+    it('subscribe should receive 1 update when team request comes first and then from user request', done => {
+      resource.subscribe('team-faster-user', mentions => {
         expect(mentions).toHaveLength(resultCraig.length + teamResults.length);
-
-        count++;
-        if (count === 2) {
-          done();
-        }
+        done();
       });
 
-      resource.subscribe('test2', mentions => {
-        expect(mentions).toHaveLength(resultCraig.length + teamResults.length);
-
-        count++;
-        if (count === 2) {
-          done();
-        }
-      });
-
-      resource.filter('craig');
+      resource.filter('team-faster-user', FULL_CONTEXT);
+      jest.runTimersToTime(REQUEST_DELAY);
     });
 
     it('subscribe should receive updates with credentials omitted', done => {
@@ -185,35 +214,51 @@ describe('TeamMentionResourceSpec', () => {
       resource.filter('esoares');
     });
 
-    it('subscribe should still receive updates when one of users or teams requests fails', done => {
-      resource.subscribe('test1', mentions => {
+    it('should still receive updates when the user request succeeds, but team request fails', done => {
+      resource.subscribe('test', mentions => {
+        // just have users result
         expect(mentions).toHaveLength(resultCr.length);
         done();
       });
-
       resource.filter('query-only-teams-fail');
+    });
 
-      resource.subscribe('test2', mentions => {
+    it('should still receive updates when the user requests fails, but team request succeeds', done => {
+      resource.subscribe('test', mentions => {
+        // just have users result
         expect(mentions).toHaveLength(teamResults.length);
         done();
       });
 
       resource.filter('query-only-users-fail');
     });
+
+    it('should show error when both user and team requests fails', done => {
+      resource.subscribe(
+        'test',
+        () => done.fail('listener should not be called'),
+        error => {
+          expect(error).toMatchObject({
+            code: 500,
+            reason: 'Internal Server Error',
+          });
+          done();
+        },
+      );
+
+      resource.filter('query-both-fail');
+    });
   });
 
   describe('#unsubscribe', () => {
-    it('subscriber should no longer called', done => {
+    it('subscriber should no longer called', () => {
       const listener = jest.fn();
-      resource.subscribe('test1', listener);
-      resource.unsubscribe('test1');
-      resource.filter('craig');
+      resource.subscribe('test123', listener);
+      resource.unsubscribe('test123');
+      resource.filter('craig', FULL_CONTEXT);
 
-      // Not desirable...
-      window.setTimeout(() => {
-        expect(listener).toHaveBeenCalledTimes(0);
-        done();
-      }, 50);
+      jest.runTimersToTime(REQUEST_DELAY);
+      expect(listener).toHaveBeenCalledTimes(0);
     });
   });