+++ packages/navigation/atlassian-switcher/examples/00-base-example.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import Button from '@atlaskit/button';
 import Drawer from '@atlaskit/drawer';
-import { Switcher, withAnalyticsLogger } from './helpers';
+import { Switcher, withAnalyticsLogger, withIntlProvider } from './helpers';
 import ErrorBoundary from '../src/components/error-boundary';
 
 class SwitcherExample extends React.Component {
@@ -36,4 +36,4 @@ class SwitcherExample extends React.Component {
   }
 }
 
-export default withAnalyticsLogger(SwitcherExample);
+export default withIntlProvider(withAnalyticsLogger(SwitcherExample));