+++ packages/core/badge/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/badge
 
+## 10.0.1
+- [patch] [98e11001ff](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/98e11001ff):
+
+  - Removes duplicate babel-runtime dependency
+
 ## 10.0.0
 - [major] [76299208e6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/76299208e6):