+++ packages/core/icon-priority/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/icon-priority
 
+## 3.0.3
+- [patch] [98e11001ff](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/98e11001ff):
+
+  - Removes duplicate babel-runtime dependency
+
 ## 3.0.2
 - Updated dependencies [9d5cc39394](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9d5cc39394):
   - @atlaskit/docs@7.0.1